requirements.txt
asgiref==3.4.1
certifi==2021.10.8
charset-normalizer==2.0.10
Django==4.0.1
djangorestframework==3.13.1
idna==3.3
psycopg2-binary==2.9.3
pytz==2021.3
requests==2.27.1
sqlparse==0.4.2
urllib3==1.26.7
